rework

+++ HT_2/task3.py
@@ -3,7 +3,7 @@
 
 sum_of_first_positive = int(input("Number: "))
 if sum_of_first_positive >= 1:
-	print(sum_of_first_positive * (sum_of_first_positive + 1) / 2)
+	print(int(sum_of_first_positive * (sum_of_first_positive + 1) / 2))
 else:
 	print(None)
